Fixed error message on first use of pid file.

+++ src/sage/server/notebook/notebook.py
@@ -1196,5 +1196,6 @@ def notebook(dir         ='sage_notebook',
     expect_quitall(verbose=False)
     from sage.misc.misc import delete_tmpfiles
     delete_tmpfiles()
-    os.remove('%s/pid'%dir)
+    if os.path.exists('%s/pid'%dir):
+        os.remove('%s/pid'%dir)
     return nb
